# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tienda Magic API",
        "description": "Documentación de la API de Tienda Magic Swagger OpenApi",
        "contact": {
            "email": "soporte@tiendamagic.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/cards": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Lista de cartas Admin",
                "description": "Obtiene una lista paginada de cartas (solo admins).",
                "operationId": "0a70ecc7dda4e1759db9126fef934979",
                "responses": {
                    "200": {
                        "description": "Lista de cartas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Crear carta manual",
                "description": "Añade una carta nueva manualmente (solo admins).",
                "operationId": "42cdcb4c431a7846921508be0edba3ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "scryfall_id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "set_code": {
                                        "type": "string"
                                    },
                                    "collector_number": {
                                        "type": "string"
                                    },
                                    "rarity": {
                                        "type": "string"
                                    },
                                    "price_eur": {
                                        "type": "number",
                                        "nullable": true
                                    },
                                    "price_usd": {
                                        "type": "number",
                                        "nullable": true
                                    },
                                    "mana_value": {
                                        "type": "integer"
                                    },
                                    "image_uri": {
                                        "type": "string",
                                        "format": "url",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Carta creada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/mod/forums/{forumId}/threads": {
            "get": {
                "tags": [
                    "Moderation"
                ],
                "summary": "Lista de hilos del foro para moderadores",
                "description": "Obtiene la lista de los hilos del foro para moderar.",
                "operationId": "8a90a33c99aacdf4c2bc1aef6e39e923",
                "parameters": [
                    {
                        "name": "forumId",
                        "in": "path",
                        "description": "ID del foro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de hilos de moderación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/mod/threads/{thread}": {
            "delete": {
                "tags": [
                    "Moderation"
                ],
                "summary": "Eliminar hilo (Moderación)",
                "description": "Elimina un hilo del foro (solo moderadores/admins).",
                "operationId": "3013f2f3dec739e64dd8fd7240a286ed",
                "parameters": [
                    {
                        "name": "thread",
                        "in": "path",
                        "description": "ID del hilo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hilo eliminado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/mod/comments/{comment}": {
            "delete": {
                "tags": [
                    "Moderation"
                ],
                "summary": "Eliminar comentario (Moderación)",
                "description": "Elimina un comentario del foro (solo moderadores/admins).",
                "operationId": "8e90bd835cd909835d66e872eda49f90",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "ID del comentario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comentario eliminado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/mod/threads/{threadId}/restore": {
            "post": {
                "tags": [
                    "Moderation"
                ],
                "summary": "Restaurar hilo (Admin)",
                "description": "Restaura un hilo que fue eliminado (solo admins).",
                "operationId": "6f32768a0d88dba61f493db6bda69764",
                "parameters": [
                    {
                        "name": "threadId",
                        "in": "path",
                        "description": "ID del hilo a restaurar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hilo restaurado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/mod/comments/{commentId}/restore": {
            "post": {
                "tags": [
                    "Moderation"
                ],
                "summary": "Restaurar comentario (Admin)",
                "description": "Restaura un comentario que fue eliminado (solo admins).",
                "operationId": "4e1bbaa7948a45d30343a667e032f6f8",
                "parameters": [
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "ID del comentario a restaurar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comentario restaurado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/wallet/recharge": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Recargar billetera",
                "description": "Inicia el flujo de recarga de saldo en la billetera virtual mediante Stripe.",
                "operationId": "20054417fdac6404f704c157e18b5ab0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "description": "Cantidad a recargar (5, 10, 20 o 50)",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sesión de Stripe Checkout creada exitosamente"
                    },
                    "422": {
                        "description": "Error de validación del monto"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/saved": {
            "get": {
                "tags": [
                    "Forums"
                ],
                "summary": "Lista de hilos guardados",
                "description": "Devuelve los hilos que el usuario ha guardado.",
                "operationId": "d282c4bd41591796d22beb72071f6101",
                "responses": {
                    "200": {
                        "description": "Lista paginada de hilos guardados"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}